+++ app/Livewire/Events/EventEdit.php
@@ -5,6 +5,7 @@ namespace App\Livewire\Events;
 use App\Livewire\Forms\EventForm;
 use App\Models\Event;
 use App\Notifications\UpcomingEvent;
+use Illuminate\Support\Facades\Log;
 use Livewire\Component;
 use NotificationChannels\WebPush\PushSubscription;
 
@@ -34,6 +35,7 @@ class EventEdit extends Component
     {
         $this->eventForm->delete();
 
+        Log::info("Event deleted", [auth()->user(), $this->eventForm->event]);
         session()->put('message', __('The event has been successfully deleted.'));
 
         return redirect($this->previousUrl);
@@ -43,6 +45,7 @@ class EventEdit extends Component
     {
         $this->eventForm->store();
 
+        Log::info("Event copied", [auth()->user(), $this->eventForm->event]);
         session()->put('message', __('The event has been successfully created.'));
 
         return redirect(route('event.edit', ['event' => $this->eventForm->event->id]));
@@ -52,6 +55,7 @@ class EventEdit extends Component
     {
         $this->eventForm->update();
 
+        Log::info("Event updated", [auth()->user(), $this->eventForm->event]);
         session()->put('message', __('The event has been successfully updated.'));
 
         return redirect($this->previousUrl);
@@ -65,6 +69,7 @@ class EventEdit extends Component
             PushSubscription::all(),
             new UpcomingEvent($this->eventForm->event)
         );
+        Log::info("Event webPush forced", [auth()->user(), $this->eventForm->event]);
     }
 
     public function render()