+++ app/Livewire/Events/EventTypeCreate.php
@@ -3,6 +3,7 @@
 namespace App\Livewire\Events;
 
 use App\Livewire\Forms\EventTypeForm;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Validation\ValidationException;
 use Livewire\Component;
 
@@ -24,6 +25,7 @@ class EventTypeCreate extends Component
     {
         $this->eventTypeForm->store();
 
+        Log::info("Event type created", [auth()->user(), $this->eventTypeForm->eventType]);
         session()->put('message', __('The event type has been successfully created.'));
 
         return redirect($this->previousUrl);