+++ app/Livewire/Events/EventTypeEdit.php
@@ -4,6 +4,7 @@ namespace App\Livewire\Events;
 
 use App\Livewire\Forms\EventTypeForm;
 use App\Models\EventType;
+use Illuminate\Support\Facades\Log;
 use Illuminate\Validation\ValidationException;
 use Livewire\Component;
 
@@ -26,6 +27,7 @@ class EventTypeEdit extends Component
     {
         $this->eventTypeForm->update();
 
+        Log::info("Event type updated", [auth()->user(), $this->eventTypeForm->eventType]);
         session()->put('message', __('The event type has been successfully updated.'));
 
         return redirect($this->previousUrl);
@@ -35,6 +37,7 @@ class EventTypeEdit extends Component
     {
         $this->eventTypeForm->delete();
 
+        Log::info("Event type deleted", [auth()->user(), $this->eventTypeForm->eventType]);
         session()->put('message', __('The event type has been successfully deleted.'));
 
         return redirect(route('event.type.index'));