+++ app/Livewire/Members/MemberGroupCreate.php
@@ -3,6 +3,7 @@
 namespace App\Livewire\Members;
 
 use App\Livewire\Forms\MemberGroupForm;
+use Illuminate\Support\Facades\Log;
 use Livewire\Component;
 
 class MemberGroupCreate extends Component
@@ -20,6 +21,7 @@ class MemberGroupCreate extends Component
     {
         $this->memberGroupForm->store();
 
+        Log::info("Member group created", [auth()->user(), $this->memberGroupForm->memberGroup]);
         session()->put('message', __('The member group has been successfully created.'));
 
         return redirect($this->previousUrl);