+++ app/Livewire/Members/MemberOverview.php
@@ -4,13 +4,18 @@ namespace App\Livewire\Members;
 
 use App\Livewire\MemberFilterTrait;
 use App\Models\Member;
+use Illuminate\Database\Eloquent\Builder;
 use Livewire\Component;
 
 class MemberOverview extends Component
 {
     use MemberFilterTrait;
 
-    public function getMembersProperty()
+    /**
+     * method to make a property available in template: $this->members
+     * @return Builder
+     */
+    public function getMembersProperty(): Builder
     {
         $memberList = Member::getAllFiltered($this->getMemberFilter());